+++ QuanLyNhanKhau/src/main/java/cnpm/quanlynhankhau/controllers/EdgeController.java
@@ -13,7 +13,7 @@ public abstract class EdgeController {
 	}
 
 	public void onNhanKhauClicked(ActionEvent event) {
-		changeScene("/cnpm/quanlynhankhau/views/QuanLyHoKhau.fxml");
+		changeScene("/cnpm/quanlynhankhau/views/QuanLyNhanKhau.fxml");
 	}
 
 	public void onHoKhauClicked(ActionEvent event) {
