+++ QuanLyNhanKhau/src/main/java/cnpm/quanlynhankhau/controllers/QuanLYNhanKhauController.java
@@ -1,0 +1,24 @@
+package cnpm.quanlynhankhau.controllers;
+
+import javafx.event.ActionEvent;
+import javafx.scene.control.TableView;
+import javafx.scene.control.TextField;
+import javafx.scene.layout.Pane;
+
+public class QuanLYNhanKhauController {
+    public Pane pnlNhanKhau;
+    public TextField tfTimKiem;
+    public TableView tvNhanKhau;
+
+    public void onThemMoiClicked(ActionEvent event) {
+    }
+
+    public void onDKTamVangClicked(ActionEvent event) {
+    }
+
+    public void onDKTamTruClicked(ActionEvent event) {
+    }
+
+    public void onKhaiTuClicked(ActionEvent event) {
+    }
+}
