+++ QuanLyNhanKhau/src/main/java/cnpm/quanlynhankhau/controllers/TrangChinhControllers.java
@@ -6,35 +6,7 @@ import javafx.scene.control.TableView;
 import javafx.scene.control.TextField;
 import javafx.scene.layout.Pane;
 
-public class TrangChinhControllers {
+public class TrangChinhControllers extends EdgeController{
     public Pane pnlTrangChu;
-    public Pane pnlNhanKhau;
-    public Pane pnlHoKhau;
-    public Pane pnlThongKe;
-    public TextField tfTimKiem;
-    public TableView tvNhanKhau;
 
-    public void onTrangChuClicked(ActionEvent event) {
-    }
-
-    public void onNhanKhauClicked(ActionEvent event) {
-    }
-
-    public void onHoKhauClicked(ActionEvent event) {
-    }
-
-    public void onThongKeClicked(ActionEvent event) {
-    }
-
-    public void onThemMoiClicked(ActionEvent event) {
-    }
-
-    public void onDKTamVangClicked(ActionEvent event) {
-    }
-
-    public void onDKTamTruClicked(ActionEvent event) {
-    }
-
-    public void onKhaiTuClicked(ActionEvent event) {
-    }
 }
