+++ QuanLyNhanKhau/src/main/resources/cnpm/quanlynhankhau/views/QuanLYNhanKhau.fxml
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.ScrollPane?>
+<?import javafx.scene.control.TableColumn?>
+<?import javafx.scene.control.TableView?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.Pane?>
+
+<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="cnpm.quanlynhankhau.controllers.QuanLYNhanKhauController">
+   <children>
+      <Pane fx:id="pnlNhanKhau" layoutX="10.0" layoutY="9.0" prefHeight="542.0" prefWidth="711.0" style="-fx-background-color: white;">
+         <children>
+            <TextField fx:id="tfTimKiem" layoutX="14.0" layoutY="34.0" prefHeight="25.0" prefWidth="283.0" />
+            <ScrollPane layoutX="14.0" layoutY="71.0" prefHeight="458.0" prefWidth="562.0">
+               <content>
+                  <TableView fx:id="tvNhanKhau" prefHeight="1230.0" prefWidth="1908.0">
+                     <columns>
+                        <TableColumn prefWidth="75.0" text="C1" />
+                        <TableColumn prefWidth="75.0" text="C2" />
+                     </columns>
+                  </TableView>
+               </content>
+            </ScrollPane>
+            <Button layoutX="605.0" layoutY="76.0" mnemonicParsing="false" onAction="#onThemMoiClicked" text="Thêm mới" />
+            <Button layoutX="601.0" layoutY="121.0" mnemonicParsing="false" onAction="#onDKTamVangClicked" text="ĐK tạm vắng" />
+            <Button layoutX="606.0" layoutY="161.0" mnemonicParsing="false" onAction="#onDKTamTruClicked" text="ĐK tạm trú" />
+            <Button layoutX="615.0" layoutY="207.0" mnemonicParsing="false" onAction="#onKhaiTuClicked" text="Khai tử" />
+         </children>
+      </Pane>
+   </children>
+</AnchorPane>
