+++ QuanLyNhanKhau/src/main/resources/cnpm/quanlynhankhau/views/Trang_chinh.fxml
@@ -3,10 +3,6 @@
 <?import javafx.geometry.Insets?>
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.Label?>
-<?import javafx.scene.control.ScrollPane?>
-<?import javafx.scene.control.TableColumn?>
-<?import javafx.scene.control.TableView?>
-<?import javafx.scene.control.TextField?>
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.Pane?>
 <?import javafx.scene.text.Font?>
@@ -66,23 +62,6 @@
                <Font size="17.0" />
             </font>
          </Label>
-      </Pane>
-      <Pane fx:id="pnlHoKhau" layoutX="-2.0" layoutY="1.0" prefHeight="538.0" prefWidth="711.0" style="-fx-background-color: blue;" />
-      <Pane fx:id="pnlThongKe" layoutX="-1.0" prefHeight="539.0" prefWidth="711.0" style="-fx-background-color: red;" />
-      <Pane fx:id="pnlNhanKhau" layoutY="-1.0" prefHeight="542.0" prefWidth="711.0" style="-fx-background-color: white;">
-         <TextField fx:id="tfTimKiem" layoutX="14.0" layoutY="34.0" prefHeight="25.0" prefWidth="283.0" />
-         <ScrollPane layoutX="14.0" layoutY="71.0" prefHeight="458.0" prefWidth="562.0">
-            <TableView fx:id="tvNhanKhau" prefHeight="1230.0" prefWidth="1908.0">
-               <columns>
-                  <TableColumn prefWidth="75.0" text="C1" />
-                  <TableColumn prefWidth="75.0" text="C2" />
-               </columns>
-            </TableView>
-         </ScrollPane>
-         <Button layoutX="605.0" layoutY="76.0" mnemonicParsing="false" onAction="#onThemMoiClicked" text="Thêm mới" />
-         <Button layoutX="601.0" layoutY="121.0" mnemonicParsing="false" onAction="#onDKTamVangClicked" text="ĐK tạm vắng" />
-         <Button layoutX="606.0" layoutY="161.0" mnemonicParsing="false" onAction="#onDKTamTruClicked" text="ĐK tạm trú" />
-         <Button layoutX="615.0" layoutY="207.0" mnemonicParsing="false" onAction="#onKhaiTuClicked" text="Khai tử" />
       </Pane>
    </AnchorPane>
    <AnchorPane layoutX="-1.0" layoutY="4.0" prefHeight="506.0" prefWidth="202.0" style="-fx-background-color: gray;" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="-1.0" AnchorPane.topAnchor="4.0">
